aad PC and PI checks Fix the Authorization bridge potential
