+++ saltapi/repository/user_repository.py
@@ -2,6 +2,7 @@
 
 import dataclasses
 from typing import List, Optional
+from xmlrpc.client import Boolean
 
 from saltapi.repository.database import database
 
@@ -87,3 +88,63 @@ WHERE u.PiptUser_Id = :user_id
         roles=[],  # TODO: get user roles
         permissions=[],  # TODO: get user permissions
     )
+
+
+async def is_user_pi(username:str, proposal_code:str) -> bool:
+    """
+    The method checks if the user is a principal investigator of a proposal.
+
+    Parameters
+    ----------
+    username
+        The PIPT username
+    proposal_code
+        The proposal code
+
+    Returns
+    -------
+    True if the user is a principal investigator else False
+
+    """
+    query = """
+SELECT Username FROM Proposal as p
+    JOIN ProposalCode as pc ON p.ProposalCode_Id = pc.ProposalCode_Id
+    JOIN ProposalContact as prc ON prc.ProposalCode_Id = pc.ProposalCode_Id
+    JOIN PiptUser as pu ON pu.PiptUser_Id = prc.Leader_Id
+WHERE Proposal_Code = :proposal_code
+    """
+    values ={"proposal_code": proposal_code}
+    result = await database.fetch_one(query=query, values=values)
+    if result[0] == username:
+        return True
+    return False
+
+
+async def is_user_pc(username:str, proposal_code:str) -> bool:
+    """
+    The method checks if the user is a principal contact of a proposal.
+
+    Parameters
+    ----------
+    username
+        The PIPT username
+    proposal_code
+        The proposal code
+
+    Returns
+    -------
+    True if the user is a principal contact else False
+
+    """
+    query = """
+SELECT Username FROM Proposal as p
+    JOIN ProposalCode as pc ON p.ProposalCode_Id = pc.ProposalCode_Id
+    JOIN ProposalContact as prc ON prc.ProposalCode_Id = pc.ProposalCode_Id
+    JOIN PiptUser as pu ON pu.PiptUser_Id = prc.Contact_Id
+WHERE Proposal_Code = :proposal_code
+    """
+    values ={"proposal_code": proposal_code}
+    result = await database.fetch_one(query=query, values=values)
+    if result[0] == username:
+        return True
+    return False
